Refactoring args utils to ensogl
